News-Pulse-Backend/requirements.txt
torch
transformers
firebase-admin
requests
pytz
python-dotenv